add phoneNumber to messages search results itrem (#1251)
